productivity_tracker/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pandas>=2.0.3
numpy>=1.21.0
python-dotenv==1.0.0
gunicorn==21.2.0
